added support for pseudo-user installation
If a pseudo-user “otpw” exists and otpw-gen is owned by it and
has SETUID set, then the hash file will be called ~otpw/john
instead of ~john/.otpw, and will be out of reach of the user.

Since we now need to read struct passwd from two users, some
memory management for struct passwd and associated strings has
become necessary, in the form of struct pwdbuf.

otpw-gen: new option -l to remove lock symlink
(useful if the lock is owned by “otpw” rather than the user)